% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{donut_data}
\alias{donut_data}
\title{Donut example}
\usage{
donut_data(n, a, b, theta)
}
\arguments{
\item{n}{Number of data points.}

\item{a}{Lower bound of the second variable.}

\item{b}{Upper bound of the second variable.}

\item{theta}{Angle of the donut.}
}
\value{
A data frame with three variables. Variable 'y' is the response,
variable 'x1' makes the donut shape with 'y', and 'x2' is a uniform random
variable between a and b. '
}
\description{
Generate data points with the shape of a donut.
}
\examples{
xy <- donut_data(n = 30, a = -1, b = 1, theta = 2 * pi)
}
