% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatgeom_group.R
\name{plot_curve.spatgeom_group}
\alias{plot_curve.spatgeom_group}
\title{Plot a \code{spatgeom_group} object}
\usage{
\method{plot_curve}{spatgeom_group}(x, type = "curve", font_size = 12, ...)
}
\arguments{
\item{x}{an object of class \code{spatgeom_group}.}

\item{type}{a string: either \code{"curve"} (default) or \code{"deriv"}.
Passed to \code{\link{plot_curve.spatgeom}} for each group.}

\item{font_size}{an integer controlling the font size. Default \code{12}.}

\item{...}{further arguments passed to \code{\link{plot_curve.spatgeom}}.}
}
\value{
A \code{cowplot} grid object (produced by
\code{cowplot::plot_grid}).
}
\description{
Plot method for objects of class \code{spatgeom_group}.
Produces a side-by-side grid of survival curve (or derivative) panels,
one column per group.
}
\examples{
set.seed(1)
xy <- donut_data(n = 60, a = -1, b = 1, theta = 2 * pi)
grp <- sample(c("A", "B", "C"), nrow(xy), replace = TRUE)
sg <- spatgeom_group(x = xy[, -1], by = grp, y = xy[, 1])
plot_curve(sg)
plot_curve(sg, type = "deriv")

}
\seealso{
\code{\link{spatgeom_group}}, \code{\link{plot_curve}}
}
