% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_geom_curves}
\alias{plot_geom_curves}
\title{plot \code{topsa} objects}
\usage{
plot_geom_curves(x, type = c("curve", "deriv"), font_size = 12)
}
\arguments{
\item{topsaObj}{an object of class \code{topsa}}

\item{nvar}{it could be a sequence from 1 to the number of variables
indicating which variables should be plotted. It could be the character
'all' for plot all the variables.}

\item{...}{further arguments passed to the \code{plot} function}
}
\value{
A plot of generated with the output of \code{topsa}. For each
variable in the model, it creates the plot of the corresponding manifold, its
symmetric reflection and its symmetric difference.
}
\description{
Plot method for objects of class \code{topsa}.
}
