% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatgeom_group.R
\name{print.spatgeom_group}
\alias{print.spatgeom_group}
\title{Print a \code{spatgeom_group} object}
\usage{
\method{print}{spatgeom_group}(x, ...)
}
\arguments{
\item{x}{an object of class \code{spatgeom_group}.}

\item{...}{further arguments passed to \code{\link{print.spatgeom}}.}
}
\value{
\code{x} invisibly.
}
\description{
Print method for objects of class \code{spatgeom_group}.
Displays a summary for each group in turn.
}
\examples{
set.seed(1)
xy <- donut_data(n = 60, a = -1, b = 1, theta = 2 * pi)
grp <- sample(c("A", "B", "C"), nrow(xy), replace = TRUE)
sg <- spatgeom_group(x = xy[, -1], by = grp, y = xy[, 1])
print(sg)

}
\seealso{
\code{\link{spatgeom_group}}, \code{\link{print.spatgeom}}
}
