% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{csr_test}
\alias{csr_test}
\title{CSR Hypothesis Testing and Global Envelope for a CSR process using
\code{spatgeom} objects}
\usage{
csr_test(
  spatgeom_obj,
  significance_level = 0.05,
  method = c("MAD", "DCLF"),
  name_ex = "ex1"
)
}
\arguments{
\item{significance_level}{a number of class numeric representing the
significance level of the test.}

\item{X}{an object of class data.frame of n dimensions.}

\item{Y}{an object of class data.frame of one dimension.}
}
\value{
a list of two objects of calls data.frame with the data of the
global envelope.  Also it saves a file of type png with the results of
the CSR Hypothesis Testing for both methods: MAD and DCLF.
}
\description{
CSR Hypothesis Testing and Global Envelope for a CSR process using
\code{spatgeom} objects
}
