% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{linear_data}
\alias{linear_data}
\title{Linear example}
\usage{
linear_data(n = 100, a = -3, b = 3)
}
\arguments{
\item{n}{Number of data points.}

\item{a, b}{Lower and upper bound of the uniform distribution.}
}
\value{
A data frame with three variables. Variable 'y = 0.6 * x1 + 0.3 * x2
\itemize{
\item 0.1 * x3' is the response, and 'x1', 'x2' and 'x3' are uniform random
variables between a and b.
}
}
\description{
Generate data points with a linear relationship.
}
\examples{
xy <- linear_data(n = 30, a = -1, b = 1)
}
