% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.spatgeom}
\alias{print.spatgeom}
\title{print a \code{spatgeom} object}
\usage{
\method{print}{spatgeom}(x, return_table = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{spatgeom}}

\item{return_table}{if \code{TRUE}, returns a data frame with the
estimated values. Otherwise, print the data frame in console. Defaults to
\code{FALSE}}

\item{...}{further arguments passed to the \code{plot} function}
}
\value{
Print the estimate given by \code{\link{spatgeom}}.
}
\description{
Print method for objects of class \code{spatgeom}.
}
\examples{

xy <- donut_data(n = 30, a = -1, b = 1, theta = 2 * pi)

estimation <- spatgeom(y = xy[, 1], x = xy[, -1])

print(estimation)

}
